pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = []
[tool.poetry.dependencies]
python-telegram-bot = "^20.6"
python-dotenv = "^1.0.1"
